Set min padding to 16.dp in swap data fields
